+++ frontend/src/router/index.ts
@@ -100,6 +100,11 @@ const router = createRouter({
       name: 'settings',
       component: () => import('../views/SettingsView.vue'),
     },
+    {
+      path: '/logs/:clustername?',
+      name: 'logs',
+      component: () => import('../views/LogsView.vue'),
+    },
   ],
 })